Move back to lotus-shed and add subcategory for indexes
